baseline/requirements.txt
torch>=2.1,<3
transformers>=4.40,<5
sentencepiece
accelerate>=0.26
pandas
tqdm
requests
gdown
llama-index-core>=0.10,<0.12
llama-index-retrievers-bm25>=0.2
